requirements.txt
streamlit==1.35.0
pandas==2.2.2
geopandas==0.14.4
shapely==2.0.4
pyproj==3.6.1
fiona==1.9.5
googlemaps==4.10
tqdm==4.66.4
openpyxl==3.1.5
fastkml>=0.12
lxml>=4.9.3
gspread==6.0.2
google-auth==2.30.0